requirements.txt
pytesseract
spacy
Pillow
flask-jwt-extended==4.5.2
python-dotenv
requests